refactor: Mover generar-auditoria.dto a auditoria-padre udistrital/sisifo_documentacion#618
