+++ src/auditoria-padre/dto/generar-auditoria.dto.ts
